requirements.txt
# Requirements untuk Pipeline Clustering Penyakit Sumatera Utara
# Core libraries untuk data processing dan machine learning

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Database connectivity
psycopg2-binary>=2.9.5
SQLAlchemy>=2.0.0

# Machine Learning & Analytics
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter & Notebook
jupyterlab>=4.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Big Data (PySpark akan diinstall via Docker image)
# pyspark>=3.4.0  # Sudah tersedia di bitnami/spark image

# Additional utilities
requests>=2.30.0
urllib3>=2.0.0
openpyxl>=3.1.0  # Untuk Excel export jika diperlukan

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
